rename: api 폴더 명 swagger와 동일하게 수정
