+++ src/utils/api/petition/getMyPetitionsByQuery.ts
