+++ src/utils/api/petition/getPetitionCount.ts
