+++ src/utils/api/petition/getPetitionsByQuery.ts
