+++ src/utils/api/petition/getStateOfAgreement.ts
